+++ docs/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Release History
 
+## 3.0.11 (2020-06-08)
+
+**Changed**
+
+- change: override variables
+    (1) testcase: session variables > step variables > config variables
+    (2) testsuite: testcase variables > config variables
+    (3) testsuite testcase variables > testcase config variables
+
+**Fixed**
+
+- fix: incorrect summary success when testcase failed
+- fix: reload to refresh previously loaded debugtalk module
+- fix: escape $$ in variable value
+
 ## 3.0.10 (2020-06-07)
 
 **Added**